Shared component for information and manifest
