+++ src/lib/ProtocolSummary/components/Stage/Items.js
@@ -1,66 +1,7 @@
-import React, { useEffect, useRef, useState } from 'react';
+import React from 'react';
 import PropTypes from 'prop-types';
 import ReactMarkdown from 'react-markdown';
-import useAssetData from '../useAssetData';
-import MiniTable from '../MiniTable';
-
-const Text = ({ content }) => <ReactMarkdown source={content} />;
-
-const Asset = ({ content }) => {
-  const {
-    url,
-    type,
-    name,
-  } = useAssetData(content);
-
-  const ref = useRef();
-  const [state, setState] = useState({ duration: 0 });
-  const metaDataListener = useRef(() => {
-    const duration = ref.current.duration.toFixed(2);
-    setState({ duration: `${duration}s` });
-  });
-
-  useEffect(() => {
-    if (ref.current && ['audio', 'video'].includes(type)) {
-      ref.current.addEventListener('loadedmetadata', metaDataListener.current);
-    }
-
-    return () => {
-      ref.current.removeEventListener('loadedmetadata', metaDataListener.current);
-    };
-  }, [ref.current, type, url]);
-
-  switch (type) {
-    case 'image':
-      return <img src={url} alt={name} />;
-    case 'video':
-      return (
-        <>
-          <video src={url} ref={ref} />
-          <MiniTable
-            rows={[
-              [<strong>Name</strong>, name],
-              [<strong>Duration</strong>, state.duration],
-            ]}
-          />
-        </>
-      );
-    case 'audio':
-      return (
-        <>
-          <audio src={url} ref={ref} />
-          <MiniTable
-            rows={[
-              [<strong>Name</strong>, name],
-              [<strong>Duration</strong>, state.duration],
-            ]}
-          />
-        </>
-      );
-    default:
-      return `"${content}":"${type}" type not supported.`;
-  }
-};
+import Asset from '../Asset';
 
 const Items = ({ items }) => {
   if (!items) { return null; }
@@ -74,13 +15,13 @@ const Items = ({ items }) => {
             case 'asset':
               return (
                 <div className="protocol-summary-stage__items-item">
-                  <Asset content={content} />
+                  <Asset id={content} />
                 </div>
               );
             default:
               return (
                 <div className="protocol-summary-stage__items-item--text">
-                  <Text content={content} />
+                  <ReactMarkdown source={content} />
                 </div>
               );
           }
