+++ src/lib/ProtocolSummary/styles/_protocol-summary-stage.scss
@@ -100,21 +100,7 @@
     }
 
     &-item {
-      @include content-box;
       margin-top: 1rem;
-
-      video,
-      img {
-        width: 50%;
-      }
-
-      img {
-        margin: 1rem 0;
-      }
-
-      video {
-        margin: 1rem 0 0;
-      }
 
       &--text {
         @include content-box;
